Move the FrontendToolsLocator to flow-server
