+++ flow-maven-plugin/src/test/java/com/vaadin/flow/plugin/common/RunnerManagerTest.java
@@ -27,6 +27,8 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.TemporaryFolder;
 
+import com.vaadin.flow.server.frontend.FrontendToolsLocator;
+
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Matchers.anyString;
