+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/FrontendToolsLocator.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.plugin.common;
+package com.vaadin.flow.server.frontend;
 
 import java.io.BufferedReader;
 import java.io.File;
